bitalino_backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
passlib
python-jose
flask
psycopg2-binary
